Always keep weights flash/blob-backed; mmap on CPU, opt-in on Linux MPU.
Remove NETKIT_WEIGHTS_IN_RAM and AOT RAM-copy paths so coefs never land in a
dedicated weight buffer. Default arenas to 64 KiB (MCU) and 64 MiB (CPU/MPU)
with CLI --arena override. Add POSIX mmap file load (MAP_PRIVATE) for CPU by
default; leave it off for MCU/MPU so RTOS and bare-metal builds match flash
buffer load, with NETKIT_MMAP=1 to opt in on embedded Linux.

+++ tools/export_mnist_cnn_int8.py
@@ -77,14 +77,14 @@ def _select_digit_cases_fast(
         label = int(y_test[i])
         if pred != label or label in used_digits:
             continue
-        # TCAS stores floats; embed prequantized int8 as float values in [-128, 127].
+        # Native int8 TCAS inputs (quantized in Python; C++ never float→int8).
         input_i8 = quantize_float_input(
             x_test[i].reshape(-1), pack.quant_layers[0].input_scale, pack.quant_layers[0].input_zero_point
         )
         cases.append(
             RegressionCase(
                 name=name_fmt.format(digit=label, i=i),
-                input=input_i8.astype(np.float32),
+                input=input_i8.astype(np.int8),
                 expected=probs,
                 label=label,
             )
